requirements-dev.txt
-r requirements.txt
mypy==0.981
PyYAML>=5.1
types-PyYAML>=5.1
pytest==6.2.5
pytest-django==4.2.0
python-dotenv==0.9.1
flake8==3.8.3
networkx==3.3
freezegun==1.1.0
black==22.3.0
